% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\title{Returns the index of an element in a vector}
\usage{
index(final, element)
}
\arguments{
\item{final}{A vector of numeric values}

\item{element}{A numeric containing the element that is to be searched for}
\value{
A numeric representing the index of the element. Returns -1 if not found.
}
\description{
Returns the index of an element in a vector.
}
