% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors.R
\name{predictors}
\alias{predictors}
\title{Returns the explanatory variable that explains the most amount of variation when fitting a linear model on.}
\usage{
predictor(df, response)
}
\arguments{
\item{df}{A data frame}

\item{response}{A string representing the response variable for comparison}
}
\value{
A string with the name of the variable that best predicts the response paramter.
}
\description{
Generates the variable that has the most amount of variation in the response variable.
}
