% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{significance}
\alias{significance}
\title{Returns the p-value of the linear model between response and explanatory}
\usage{
significance(df, response, explanatory)
}
\arguments{
\item{df, }{data frame}

\item{response, }{response variable}
}
\value{
p-value of the linear model
}
\description{
Returns the p-value of the linear model
}
