% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\title{Returns the index of element in vector}
\usage{
index(finals, element)
}
\arguments{
\item{finals, }{vector of sums of squares for each explanatory variable correlation}

\item{element, }{element to be searched for}
}
\value{
index of elementi n vector
}
\description{
Returns the index of element in vector
}
