% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors.R
\name{predictors}
\alias{predictors}
\title{Generates the predictors}
\usage{
predictors(df, response)
}
\arguments{
\item{df}{dataframe}
}
\value{
Sum of squares for the variable that explains the most variation in the dataset
}
\description{
Generates the predictors
}
